pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hunter-backend"
version = "0.1.0"
description = "Production scraping backend for Polish real estate (5 sources, normalization, Supabase)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "supabase>=2.0.0",
    "loguru>=0.7.0",
    "playwright>=1.40.0",
    "apscheduler>=3.10.0",
    "pytz>=2024.1",
    "pyyaml>=6.0",
    "flask>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21"]

[project.scripts]
hunter = "hunter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
